KAFKA-19800-2: Moved NetworkPartitionMetadataClient to group-coordinator module
